requirements.txt
tensorflow_hub==0.11.0
tensorflow==2.3.1
streamlit==0.78.0
numpy==1.17.4
matplotlib==3.3.3
ipython
Pillow==8.1.2
